Module Bitstring_lib

module Bitstring : sig ... end

The Bitstring_lib module provides simple utilities for working with lists used to represent sequences of bits or other elements. It includes different orderings (endianness) and padding utilities.