Module Bitstring_lib.Bitstring

The Bitstring_lib module provides simple utilities for working with lists used to represent sequences of bits or other elements. It includes different orderings (endianness) and padding utilities.

module type S = sig ... end

module Msb_first : sig ... end

Msb_first treats lists as having the most significant element at the head. Padding is added to the right end of the list.

module Lsb_first : sig ... end

Lsb_first treats lists as having the least significant element at the head. Padding is added to the left end of the list.

val pad_to_triple_list : default:'a -> 'a list -> 'a Tuple_lib.Triple.t list

pad_to_triple_list ~default xs converts a list xs to a list of 3-tuples. If xs does not divide evenly into triples, missing values are filled in with default.